+++ Broker_WorldQuests/Globals.lua
@@ -35,6 +35,12 @@ BWQ.mapID = 0
 -- /run local mapID = C_Map.GetBestMapForUnit("player"); print(format("You are in %s (%d)", C_Map.GetMapInfo(mapID).name, mapID))
 -- Also, see https://www.wowhead.com/guide/list-of-zone-ids-for-navigation-in-wow-and-tomtom-19501
 BWQ.MAP_ZONES = {
+	[CONSTANTS.EXPANSIONS.MIDNIGHT] = {
+		[2395] = { id = 2395, name = C_Map.GetMapInfo(2395).name, quests = {}, buttons = {}, }, -- Eversong Woods 12.0
+		[2437] = { id = 2437, name = C_Map.GetMapInfo(2437).name, quests = {}, buttons = {}, }, -- Zul'Aman 12.0
+		[2413] = { id = 2413, name = C_Map.GetMapInfo(2413).name, quests = {}, buttons = {}, }, -- Harandar 12.0
+		[2405] = { id = 2405, name = C_Map.GetMapInfo(2405).name, quests = {}, buttons = {}, }, -- Voidstorm 12.0
+	},
 	[CONSTANTS.EXPANSIONS.THEWARWITHIN] = {
 		[2248] = { id = 2248, name = C_Map.GetMapInfo(2248).name, quests = {}, buttons = {}, }, -- Isle of Dorn 11.0
 		[2214] = { id = 2214, name = C_Map.GetMapInfo(2214).name, quests = {}, buttons = {}, }, -- The Ringing Deeps 11.0
@@ -97,6 +103,7 @@ BWQ.MAP_ZONES = {
 -- The following table is used to sort the zones when displayed. This table should only include zones that are in the 
 -- BWQ.MAP_ZONES table above.
 BWQ.MAP_ZONES_SORT = {
+	[CONSTANTS.EXPANSIONS.MIDNIGHT] = 		{	2395, 2437, 2413, 2405												},
 	[CONSTANTS.EXPANSIONS.THEWARWITHIN] = 	{	2248, 2214, 2215, 2255, 2213, 2369, 2346, 2371, 2472, 241			},
 	[CONSTANTS.EXPANSIONS.DRAGONFLIGHT] = 	{	2022, 2023, 2024, 2025, 2085, 2151, 2133, 2200						},
 	[CONSTANTS.EXPANSIONS.SHADOWLANDS] =  	{	1525, 1533, 1536, 1565, 1543, 1970									},
@@ -108,6 +115,7 @@ BWQ.MAP_ZONES_SORT = {
 -- are no world quests in Dornogal, we want for the addon to recognize that it is a TWW zone so that it switches to the 
 -- THEWARWITHIN expansion properly.
 BWQ.MAP_ZONES_EXTRA = {
+	[CONSTANTS.EXPANSIONS.MIDNIGHT] = 		{	2393, 2424, 2576, 2536, 2541										},
 	[CONSTANTS.EXPANSIONS.THEWARWITHIN] = 	{	2339, 2328, 2216, 2255												},
 	[CONSTANTS.EXPANSIONS.DRAGONFLIGHT] = 	{	2112, 2239															},
 	[CONSTANTS.EXPANSIONS.SHADOWLANDS] = 	{	1699, 1550, 1670, 1671, 1961, 1701, 1698, 1707, 1911, 1912			},